Added safety checks and updated README
Added checks to see if legend file exists and if the new directory for writing resultant outputs could be created and exists. Increased max. possible dimensionality to 5 - assigned position 5 of the dimensionality vector for representing whether the dataset is a raster, vector or neither. (In the future a redundancy check could be kept based on this - e.g. positions 1 and 2 can't be 0 if position 5 is 0 or 1 etc.)

 Updated README by adding a general description of the package and a sub-section on the (draft) strategy being followed showing how this package fits into the overall architecture and workflow.