Modified relative path of filenames in legend for output
Modified the filenames of the files as output after running the UDF. Now the filename in the output legend is relative to the path of the legend file itself.

Also added example input and output for UDFs from and to backends implementing a draft naming convention and directory structure. A draft structure of a sample R UDF (a R script file to be uploaded by the user in the backend) is also uploaded.